+++ sw_listener/src/apis.rs
@@ -1,37 +1,47 @@
-use crate::hashmap::HASHMAP;
-use actix_web::{post, web, App, HttpResponse, HttpServer, Responder};
+use crate::hashmap::{QUICMAP, TCPMAP};
+use crate::quic::handle_stream;
+use actix_web::{delete, post, web, App, HttpResponse, HttpServer, Responder};
+use log::info;
 use serde::{Deserialize, Serialize};
-use tokio::io::{AsyncReadExt, AsyncWriteExt};
-use tokio::net::{TcpListener, TcpStream};
+use tokio::net::TcpListener;
 
 #[derive(Debug, Serialize, Deserialize)]
-struct PostObj {
+struct OpenObj {
   uid: String,
   port: u16,
   connect_address: String,
   connect_port: u16,
 }
 
+#[derive(Debug, Serialize, Deserialize)]
+struct CloseObj {
+  port: u16,
+}
+
 #[post("/open")]
-async fn open(json: web::Json<PostObj>) -> impl Responder {
-  let map = HASHMAP.lock().await;
-  if map.contains_key(&json.uid) {
-    match TcpListener::bind(("127.0.0.1", json.port)).await {
+async fn open(json: web::Json<OpenObj>) -> impl Responder {
+  let quicmap = QUICMAP.lock().await;
+  let max_vector_size: usize = 1024;
+  info!("opening port {}", json.port);
+  if quicmap.contains_key(&json.uid) {
+    match TcpListener::bind(("0.0.0.0", json.port)).await {
       Ok(listener) => {
-        println!("TcpListener created successfully!");
+        let mut tcpmap = TCPMAP.lock().await;
+        tcpmap.insert(json.port, listener);
+        info!("TcpListener created successfully!");
         tokio::spawn(async move {
           loop {
-            let (stream, _) = listener.accept().await.unwrap();
-            println!("Accepted connection from: {:?}", stream.peer_addr());
+            let (stream, _) = tcpmap.get(&json.port).unwrap().accept().await.unwrap();
+            info!("Accepted connection from: {:?}", stream.peer_addr());
             let addr = format!("{}:{}", &json.connect_address, &json.connect_port.to_string());
-            handle_stream(stream, 1024, &json.uid, addr).await;
+            handle_stream(stream, max_vector_size, &json.uid, addr).await;
           }
         });
         HttpResponse::Ok().body("TcpListener created successfully!")
       }
       Err(e) => {
-        println!("Failed to create TcpListener: {}", e);
-        HttpResponse::InternalServerError().body("Failed to create TcpListener")
+        let body = format!("Failed to create TcpListener: {}", e);
+        HttpResponse::InternalServerError().body(body)
       }
     }
   } else {
@@ -39,73 +49,19 @@ async fn open(json: web::Json<PostObj>) -> impl Responder {
   }
 }
 
-async fn handle_stream(mut manager_stream: TcpStream, max_vector_size: usize, uid: &String, connect_addrs: String) {
-  let map = HASHMAP.lock().await;
-  let connection = map.get(uid).unwrap();
-
-  // got SendStream and RecvStream
-  let (mut send, mut recv) = connection.open_bi().await.unwrap();
-
-  tokio::spawn(async move {
-    loop {
-      let mut buf1 = vec![0; max_vector_size];
-      let mut buf2 = vec![0; max_vector_size];
-
-      //
-      // FC HELLO (share edge configuration)
-      //
-      send.write_all(connect_addrs.as_bytes()).await.unwrap();
-      send.write_all(&buf1[0..max_vector_size - connect_addrs.as_bytes().len()]).await.unwrap();
-      println!("FC HELLO to sw_connector with edge conf: {}", connect_addrs);
-
-      //
-      // stream to stream copy loop
-      //
-      loop {
-        tokio::select! {
-          n = recv.read(&mut buf1) => {
-            match n {
-              Ok(None) => {
-                println!("local server read None ... break");
-                break;
-              },
-              Ok(n) => {
-                let n1 = n.unwrap();
-                println!("local server {} bytes >>> manager_stream", n1);
-                manager_stream.write_all(&buf1[0..n1]).await.unwrap();
-              },
-              Err(e) => {
-                eprintln!("manager stream failed to read from socket; err = {:?}", e);
-                break;
-              },
-            };
-            println!("  ... local server read done");
-          }
-          n = manager_stream.read(&mut buf2) => {
-            println!("manager client read ...");
-            match n {
-              Ok(0) => {
-                println!("manager server read 0 ... break");
-                break;
-              },
-              Ok(n) => {
-                println!("manager client {} bytes >>> local server",n);
-                send.write_all(&buf2[0..n]).await.unwrap();
-              },
-              Err(e) => {
-                eprintln!("local server stream failed to read from socket; err = {:?}", e);
-                break;
-              }
-            };
-            println!("  ... manager read done");
-          }
-        };
-      }
-    }
-  });
+#[delete("/close")]
+async fn close(json: web::Json<CloseObj>) -> impl Responder {
+  let mut tcpmap = TCPMAP.lock().await;
+  if let Some(listener) = tcpmap.remove(&json.port) {
+    let _ = listener; //listenerの所有権を解放
+    HttpResponse::Ok().body("TcpListener closed successfully!")
+  } else {
+    HttpResponse::InternalServerError().body("No TcpListener exists for the specified Port.")
+  }
 }
 
 pub async fn create_app(addr: &str, port: u16) -> () {
+  info!("API listening on {}:{}", addr, port.to_string());
   let app = || App::new().service(open);
   HttpServer::new(app).bind((addr, port)).expect("Can not bind").run().await.expect("Server failed");
 }